fix(api): cal check: terminal drop tips during delete (#5664)
We always want to drop tips in the trash when we end sessions - whether
we do a smooth exit or not (and perhaps whether another client hops in
or not). We also want those drop tips to happen when we actually exit
the session, not right before the session complete screen (which is
triggered by the checkComplete state).

To accomplish both these goals, remove the state transition hooks when
moving to trashComplete that trigger the drop tip and rely on the
behavior in the session deleter (and refactor it a bit to avoid spurious
moves, since previously you would move around some before _trash_tip
realized there was nothing to drop).