refactor(api): split geometry.plan_moves (#5662)
opentrons.protocol_api.geometry.plan_moves is a helper function
responsible for planning arc moves that always go high enough to avoid
labware, possibly avoid thermocyclers, and switch critical points. It did all of
- Understand the deck enough to know what "high enough" means, including
constraints like "not too high for the instrument"
- Understand protocol api semantics enough to know when an arc move is
actually necessary, and issue direct moves if possible
- Understand the deck enough to know how to dodge thermocyclers
- Understand hardware enough to switch critical points
- Understand labware semantics enough to figure out which critical
points to use

It was also long enough to need a mccabe exception in the linter.

Because it was in geometry, anything in the rest of the system that
wanted to move in an arc either needed to import the protocol api
subsystem or roll its own arcs.

This splits plan_moves up into three components:
- plan_moves is a one-size-fits-all entrypoint for callers that do know
all about decks and everything (like the protocol api). It calls the
other components and still does all of the above... but with a lot of
work done in helpers to avoid the complexity. The parts it does on its
own are deciding if avoiding the thermocycler is necessary, and issuing
direct moves if possible, and figuring out which critical points to use
- geometry.safe_height implements the "understand the deck enough to
figure out a safe height" part of the stack, which can now be called
independently without involving all the rest of the stuff
- hardware_control.util.plan_arc, which takes an origin point + cp, a
destination point + cp, and a z height and does the part of actually
generating the moves and blending between the two requested critical
points. It also takes a list of internal waypoints, which
geometry.plan_moves provides for dodging the thermocycler

The API to the protocol api is unchanged, but now other components of
the system can build their own arc moves with similar logic but choosing
their own components. For instance (and in an upcoming PR), deck
calibration check can use these components to build arc moves with
proper z heights (using safe_height()) but specifying all critical
points (using hardware_control.util.plan_arc() directly without using
geometry.plan_moves).