fix(robot-server,api): bug in deck calibration
Root cause is that the field in api was a str but the fastapi passed in an enum. The str is a key in a dict and the lookup using the enum failed due to a str conversion before lookup.
Solution is to move the enum into the API and have it referenced by fastapi model. Also, remove the str conversion before look up.

closes #5688