chore(api): add util to plot session state machines
This will parse the calibration check session (and eventually later
sessions) and produce a nice pdf graph if you have graphviz installed.
This can be useful for, for instance, noticing if there's a state that's
never used because it has no edge (hint hint).

To use, update your pipenv and install graphviz from somewhere (on
linux, your package manager; on OSX, homebrew; on windows, who knows)
and make sure its `dot` utility is on your path.