fix(api): proper hardware controller teardown
opentrons_execute was printing an exception when it stopped, because we
weren't properly tearing down the module watcher and were relying on the
interpreter to do so - not great. We now explicitly tear down the
hardware control stack.

When we tear down the hardware controller, we cancel some running tasks.
When you cancel a running task, it pends a CancelledException, but the
task needs to execute on the next loop spin, so we need to spin the
loop. This is the complex annoying logic in
thread_manager.__getattribute__.

Closes #5728