Adds multiplication operator(s) to types.Point
This allows scaling an offset by a fixed amount, making it easy to combine e.g. a direction vector and a distance.

Previously:
Point(1,2,3)*2 => (1, 2, 3, 1, 2, 3)

Feel totally free to reject this e.g. if Points are not meant to be used for offset vectors. I couldn't find tests for this but feel free to point me (NPI) to them and I can add one.