refactor(api): raise GPIO warning, and add detail (#5782)
This raises the level of the warning about the failure to initialise a GPIO device from `info` to `warning`, making it more likely that it will be displayed (e.g. in Jupyter). Without this there is no indication given to the user of e.g. a Jupyter notebook of why their commands to control lights are doing nothing. It also adds some detail about what functionality is likely to be affected.