refactor(api, shared-data): move py shared data accessors (#5777)
The JS stack has integrated type definitions and bindings for accessing the content in shared_data. Now the Python stack does too.

This adds a new python package, opentrons_shared_data, which will be built and uploaded to pypi with the same rules as the api package. It is loaded into the api and robot server pipenv spaces editably, and the api package marks it as a dependency to pypi.

opentrons_shared_data has little code - most of it is slightly-customized accessors for the json definitions - but it has a lot of types, types that define the structure of the shared data files and are ensured to be correct using the runtime type checker package typeguard in the unit tests (there is still no runtime type checking in the production code).

In addition, the api package and tests have been modified to use the types and loaders in opentrons_shared_data instead of directly accessing the content.
New packa
shared data now lives in opentrons_shared_data, with accessors identical
to opentrons.system.shared_data.

Also, add py.typed to the opentrons package so mypy can typecheck it externally.

Co-authored-by: Brian Arthur Cooper <brian@opentrons.com>