Removed many database migration print statements (#5847)
This removes the database migration statements that are printed every time the robot module is imported, which are confusing for users, in favour of `log.info` messages. It leaves print statements intact when migration actually happens so that if the interface does not lock up without any information on progress.