fix(build): allow test pypi uploads to fail again
The shared-data pr had a small makefile refactor to coalesce the deploy
and deploy-staging python makefile tasks, but what that forgot is that
uploads to test pypi must be allowed to fail, since every edge build
in the same version will upload a wheel with the same name, which pypi
doesn't allow.

This adds that functionality back to the python_upload_package makefile
function

It also adds a bit of complexity to allow callers to specify a prefix
command since the shared-data task needs to execute in a subdirectory.