chore(build): py: add .devN tag for edge builds
These get uploaded to test pypi (where they are by and large never
downloaded, but it allows us to dry-run uploads to production pypi).
Test pypi doesn't allow you to reuse filenames, just like real pypi, and
each edge build in the same version will have the same wheel filename.
In the past, we overcame this by allowing the test pypi upload to fail,
but then... that doesn't really test anything.

Luckily, PEP440[1] which defines python versioning conventions, allows a
.devN tag to be at the end of versions, with the intent being that the N
is a build number. So this pr adds the capability to specify a
BUILD_NUMBER= environment variable - which travis edge builds now do -
which will get added to
- the wheel filenames
- the actual versions encoded in the wheel metadata
- the required version for opentrons_shared_data in opentrons

and thus allow us to upload to pypi on every edge build.