fix(api): do not allow v1 imports in v2 protocols
If you import v1 protocols in your v2 protocol, it slows things down and
kinda messes them up since those modules are designed to do a lot of
work on import, and it's work that actively shouldn't be done unless
necessary.

If you then _use_ those modules, it can play absolute havoc with
assumptions inside the protocol session machinery.

This is just straight up not supported, and this pr adds an error that
will occur when you try and upload a protocol that does this.

Closes #5852