refactor(app): calcheck: confirm exit modal
Add a modal that appears when you click on exit (in most circumstances)
during calibration check that prompts you to confirm the exit.

If the exit button is from a terminal screen, there is no confirmation.

In addition, hide the title bar exit button when it isn't necessary,
most notably when the spinner is displayed.

Note that the exit confirmation should actually lead to a generic
summary screen and not immediately exit, but that screen doesn't exist yet.