chore(api): fix lots of Sphinx warnings (#5794)
* Fix "Unexpected indent"

* Fix "Explicit markup ends without a blank line"

Some of the reference documentation did this:

```
Examples
--------
..
>>> foo()
```

This warning is about the `..` on its own line.

Although those `..`s have been there since 2016 (e.g. 9c7af54de5ce096479a018435e1d514537b7c80e), I'm pretty sure they're just a misunderstanding of reST syntax?

In reStructuredText, `..` is the beginning of an [explicit markup block][1].  But it's usually (always?) followed by the name of a directive, e.g. `.. hint::`.  It's also supposed to be followed by an indented block, which is terminated by a blank line.  None of that happens here.

Lines beginning with `>>>` are [doctest blocks][2], which Sphinx picks up independently.  They're unaffected by the preceding `..`.

(Also, the "Examples" section is treated specially our [the Napoleon extension][3].  But I don't think that has any bearing here.)

So, we just remove all the errant `..` lines.

[1]: https://docutils.sourceforge.io/docs/ref/rst/restructuredtext.html#explicit-markup-blocks
[2]: https://www.sphinx-doc.org/en/master/usage/restructuredtext/basics.html#doctest-blocks
[3]: https://www.sphinx-doc.org/en/master/usage/extensions/napoleon.html

* Fix more "Explicit markup ends without a blank line"

This time, it's just a missing blank line.

* Fix "Title underline too short"

* Fix "Content block expected for the 'note' directive; none found"

* Remove empty note admonition block

* Fix "Could not lex literal_block as 'python'"

The syntax highlighter has trouble here because of the Jupyter Notebook syntax.  Since it falls back to skipping syntax highlighting entirely, we don't lose anything by just disabling it here.

* Fix "Duplicate explicit target name"

See: https://stackoverflow.com/questions/27420317/restructured-text-rst-http-links-underscore-vs-use

* Fix "Inline interpreted...reference start-string without end-string"

* Fix syntax of "new in version" markers for params

The .. versionadded markers in these param lists had wrong indentation.
In one case, this caused the list to be split up when it was rendered.

Fix this by using a method-level `.. versionadded` instead of param-level ones.

We could also have fixed the indentation to get the param-level markers to work.  But, that would have rendered with slightly uglier formatting.  See sphinx-doc/sphinx#4227.

* Fix "Field list ends without a blank line"

* Fix broken cross-references using the wrong cross-reference type

Lots of places used :attr: (attribute) to try to cross-reference to something that looked like an attribute, but in fact was a property, i.e. a function.

To make these cross-references less brittle, replace them with the generic :obj:.

* Bonus: Fix stray "..." breaking code block parsing

In the rendered output, this made the block not use code formatting.

* Bonus: Fix wrong emphasis syntax

This was being rendered with literal underscores.

* Lint fixup: Remove whitespace from blank lines