refactor(robot-server, app): wire up tip length calibration session details and commands (#6132)
Add contents to the session details returned from the robot-server upon performing an action within
a tip length calibration session. Replace the stubbed out calls to commands and session in the run
app, with actual calls and references to the created session through sessions api machinery in
redux. Further divide the session concerns of the tip length calibration from the page component
within which it is rendered.