refactor(api): load new tip length calibration if ff set
If the robot-side feature flag use_tip_length_calibration is set, we
will use the new tip length calibration data in the protocol api and the
hardware controller (so, APIv2 only for now).

Loading tip length calibration data requires knowing about both the
pipette id and the labware, which means it's not a natural fit for just
the opentrons.protocol_api.labware module. The place where both of those
data are present is in the InstrumentContext, which features a
tip_length_for method used by everything that cares about tip length
that is a perfect place to add the loading.

In addition to the instrument_context change, the new tip length data is
actually used slightly differently. Previously, tip length calibration
was folded into the instrument offset in the z coordinate there. That
means that when applying the instrument offset to critical point
computation, applying the z or not depends on whether there's a tip.

Now that tip length calibration is separate, that's not true, so if the
feature flag is on the entire instrument offset can be added to the
critical point.

Unfortunately, this means things won't work quite right when you toggle
on the feature flag unless you zero out the instrument offset z part.

Closes #6168