fix(api): Clarify what is behind GPIO connection problems (#6204)
(Only one device can connect at a time).

The message displayed when we fail to connect to GPIOs in Jupyter/Python script because another process is already using them doesn't explain that the reason stopping the server works is that it frees up the GPIOs for connecting. This may (or may not) be behind for example #6164.

This adds that to the message.