fix(robot-server): Log uncaught exceptions
While uncaught exceptions that bubble all the way up to the FastAPI
layer all definitionally bugs and shouldn't be expected in a running
server, we do want to be able to log them and provide nice JSONApi error
results if they do occur. This adds
- An ErrorMiddleware handler to catch all exceptions
- Logs it
- Returns a json error page

It also enables logging for other internal server components.