fix(robot-server, app): fix broken transition from chained commands in tip length cal
Within the tip length calibration user flow, when the user save's the offset of their pipette's
nozzle the api receives two commands in quick succession. The first one tells the robot to save that
offset, and the second tells the robot to move to the tip rack in preparation for picking up a tip.
The latter command was only triggerable from the resulting state of the successful transition
triggered by the "saveOffset" command. Because of the async nature of the api, this meant that the
saveOffset command might not have completed the transition to the state that has "moveToTiprack" as
a valid trigger. This fix avoids the race condition, by only transitioning to the next state from
the "moveToTiprack" command