fix(api): set parent for in-slot offsets to empty
We store calibration offsets by parent - including by slot. When an
offset is stored for a slot, the parent key (which is used to determine
the location of the calibration file) is saved as the labware hash.

However, this is not how it should be expressed when loaded - the parent
of a slot-relative offset is empty until and unless we start saving
different calibration offsets for the same labware in different slots.

Also, fill out the example fields for the labware calibration model in
the robot server and restore an early return in get labware calibrations
if there are no calibrations to return.