refactor(robot-server): tlc: clients specify tiprack def
TipLengthCalibration needs to happen for the tiprack that a particular
protocol is using. It also needs to happen for selectable tipracks to
prepare for instrument offset calibration. To support both these
workflows, we now require the client to send the labware definition of
the tiprack to calibrate in the create args for the session.