refactor(app): provide tiprack def to tip length cal
Now that the tip length calibration flow needs a tiprack labware
definition, we can pull that definition from the protocol and provide
it. This also requires altering various components to properly render
arbitrary labware if there is no pre-cooked image.

This also copies over the rest of the tiprack images we have for
non-opentrons labware from labware-library. These should be moved to
shared-data.