refactor(app): persist whether tip cal user has cal block via config (#6223)
Upon starting tip length calibration, the user is asked if they have access to a calibration block.
If they check the box in the UI to "not show this message again", then their selection will be
persisted to an app advanced setting in config. Because this setting can be "undefined" for those
who haven't see the prompt or have clicked through without checking the box, the toggle on the
settings page will only be present after initially checking the box in the pre-tip calibration
modal. At which point they should never see the modal again.