refactor(robot-server,app): Use tiprack from protocol in tip length calibration (#6242)
Tip length calibration must be run on the same tiprack that is used in the protocol. It also in the future will have to be run on either a standard or user-selectable tiprack to establish tip length before instrument offset; and in the farther future, tip length calibration may be runnable at will outside of the protocol flow.

To support this,

- The tip rack definition to calibrate is now sent as a createParam to the tip length calibration session
- The app loads the tip rack definition for the tip rack to calibrate using the same mechanisms it uses for displaying labware. This unfortunately means #6234 affects the app as well, but it handles custom labware sourced in json protocols or protocols uploaded from the local app acceptably.

Changelog:

robot-server: tlc: clients specify tiprack def

TipLengthCalibration needs to happen for the tiprack that a particular
protocol is using. It also needs to happen for selectable tipracks to
prepare for instrument offset calibration. To support both these
workflows, we now require the client to send the labware definition of
the tiprack to calibrate in the create args for the session.

app: provide tiprack def to tip length cal

Now that the tip length calibration flow needs a tiprack labware
definition, we can pull that definition from the protocol and provide
it. This also requires altering various components to properly render
arbitrary labware if there is no pre-cooked image.

This also copies over the rest of the tiprack images we have for
non-opentrons labware from labware-library. These should be moved to
shared-data.

Closes #6232 

Co-authored-by: Brian Arthur Cooper <brian@opentrons.com>