fix(app): don't doublecount labware on modules
If you have a labware on a module, and another of that labware on the
deck, the app will display (x2) next to the entry for the labware on the
deck in the file info panel. This is becaues of too-loose matching on
parent in a new selector. Tighten up that matching so labware on modules
doesn't get doublecounted.