refactor(api): expose apilevel-based flow rates in hc
Expose the dicts mapping api level to default flow rates for pipettes to
hardware controller consumers, and refactor the pipette flow rate
tracking to match.

Because the dicts exposed in the attached_instruments calls are getting
unwieldy, add a TypedDict definition for them; to remove the verbosity
of the key checking in tests, and to make sure the type stays synced
with reality, add typeguard checking so we can actually rely on the
types.

Move flow rates to something tracked by Pipette class instances in
instance members rather than modifying the config. This requirs
modifying the hardware controller api to access these values through the
new property accessors instead of by poking into the config.

Unfortunately this requires a slight leakage of api level concepts into
the hardware controller, but only to the extent of using the hardcoded
2.0 flow rates by default.