fix(api): don't stomp other loggers in log config
A recent robot-server PR deferred API loading until after the server is
up, which is a great change but means that the API-side logging
configuration now stomps on the robot server's logging configuration, so
robot server messages are not logged.

This commit removes the stomping and the robot server's logs should now
be visible again.