fix(api): avoid hitting endstop during retract fast home
The smoothie fast home command moves to near the home position at normal
speed, swallowing errors, and then issues a home to minimize the amount
of time a possibly-slower home must occur.

However, even though the error is swallowed at the level of the
fast_home, if the move triggers the endstop switch the smoothie would
issue a home before returning the error.

While it's not generally safe to mess around with the retract distance,
in the case of pick up tip - the most common case for retracts after
expected skips - we can confidently add a couple millimeters to the
retract to avoid it.

This should also account for the extra distance moved (and possibly
skipped) for pipettes that have more than one press of pick up tip.