refactor(app, robot-server): add inspecting tip state to TLC, remove all stubs
In order to better match the designed flow, this separates out the inspecting tip state of the user
flow from the preparing pipette state. This prevents a bug where the presence of the loading spinner
between those two steps clears the view layer reference to whether or not the pipette has attempted
a pick up. That value should be an implicit part of the user flow. This branch also replaces all
remaining stubbed data in the app side with the correct value from the tip length calibration
session.