fix(api): Save v2 labware in api v1 with URIs (#6325)
Not enough information was being saved in api v1 for new labware definitions in order for the calibration to carry over properly to api v2 protocols/be displayed in the app.

Instead, 
- Get the uri for the labware definition
- pass in the labware definition to the save function
- Make the uri details function slightly more robust so it won't fail if no URI exists for some reason. It will default to returning the entire thing as a loadname.

Limitations:

There is not an easy way for us to get the proper names of the temp deck + mag deck to save the calibrations accordingly.

That means that any calibrations saved in api v1 for modules will not translate to api v2, and also will not show up in the app for api v1.

An unrelated limitation is on the temperature module GEN1 vs GEN2. Although the calibrations should be the same, the system is saving them as two separate calibrations. We will document this to hopefully prevent confusion.