fix(app, app-shell): use mtime for custom labware date display
The custom labware screen in the more tab currently displays the create
time of each custom labware file, using node fs.birthtimeMs. This has a
couple problems, like
- it may use the date the file was created originally (e.g. passing it
around in a zip file may save the original file metadata, on some
filesystems)
- since we update labware by overwriting the file, on some operating
systems or filesystems this may actually overwrite the contents of the
file rather than deleting and creating a new file, which would leave the
creation time unchanged.

The second is particularly pernicious, since the time displayed in the
labware card is the only way the user can tell the file has changed.

We can solve this by using the file modification time.

Closes #6381