fix(robot-server): make systemd startup more tolerant
We see occasional problems where the robot server may be crashing (it's
hard to debug since most of the time users don't remember their robot's
ip, and the app can't display ip addresses until the robot server is
up). This bumps the timeout before the robot server is considered failed
to 3 minutes, which should be sufficient for whatever is going wrong;
and it also commands a restart if the unit fails, just in case.