refactor(api): cache_instruments doesn't reconfigure same pipettes
cache_instruments used to let you specify names or models, and this is a
pain and is awful and has no actual valid use case. The hardware setup
of the simulator should only be specified at init time; the use case of
passing things to cache_instruments is
1. reconfigure the simulator, if you want, to return what you asked for,
in general (i.e. by name)
2. verify in the simulator or hardware controller that the hardware
matches what you expect

It is not to just pass in whatever and have more or less the right thing
happen.

By changing this and actually typing these functions, it lets us push
pipette config loading down to backends, which makes it much easier to
check if the pipette we're currently querying is the same as the pipette
we used to know was attached. If it is (and there's no backcompat
messing around) then congrats, we don't actually have to reconfigure
things.