refactor(api): Add a method to reset instrument state
This adds the "clear internal state" semantics that cache_instruments
used to have to a new method called reset_instruments in the hardware
control API. This can be called when necessary to make sure the state is
clean without incurring a smoothie-level reconfiguration (i.e. resetting
the unstick timer). It should only be called in server-level
orchestration components before and after protocol runs, not inside
anything reachable from protocols.