refactor(api,robot-server): use typing-extensions at runtime
We consistently have bugs about accidentally requiring typing-extensions
at runtime. As we use it more and more, it's also less and less
reasonable to keep it out of the runtime environment. With the package
added to buildroot in https://github.com/Opentrons/buildroot/pull/101,
let's just select it as a runtime dependency and free ourselves from
having to maintain these separate annoying dependency tracks.