chore(build): add pr workflows for shared-data and robot-server
These work very similarly to the api pull request jobs. They have
pull_request_target triggers as well as push triggers for the workflow
definitions themselves.

The robot-server workflow checks for changes to the robot-server itself
and to its dependencies and transitive dependencies (api, shared-data
definitions and python bindings).

The shared-data workflow will be the first that also covers js; for now,
it only does the python binding lints and tests.