refactor(api): hc: reconfigure pipettes only when they change (#6456)
Changing semantics of cache_instruments:

cache_instruments used to let you specify names or models, and this is a
pain and is awful and has no actual valid use case. The hardware setup
of the simulator should only be specified at init time; the use case of
passing things to cache_instruments is
1. reconfigure the simulator, if you want, to return what you asked for,
in general (i.e. by name)
2. verify in the simulator or hardware controller that the hardware
matches what you expect

It is not to just pass in whatever and have more or less the right thing
happen.

By changing this and actually typing these functions, it lets us push
pipette config loading down to backends, which makes it much easier to
check if the pipette we're currently querying is the same as the pipette
we used to know was attached. If it is (and there's no backcompat
messing around) then congrats, we don't actually have to reconfigure
things.

Adding internal state clearing semantics:

This adds the "clear internal state" semantics that cache_instruments
used to have to a new method called reset_instruments in the hardware
control API. This can be called when necessary to make sure the state is
clean without incurring a smoothie-level reconfiguration (i.e. resetting
the unstick timer). It should only be called in server-level
orchestration components before and after protocol runs, not inside
anything reachable from protocols.

Internal refactors:

Moving some of the more complex pipette config detection logic to the 
pipettes module cleans things up and makes sense conceptually.

We also now only do hardware reconfiguration if absolutely necessary,
and the infrastructure and types required for that lets us remove the last
place we were touching the smoothie driver in the hardware control api
directly.