fix(api): limit calibration jogs when going upward (#6565)
Changelog:
- better optional bounds checking for _move

Refactor the bounds checking done in _move to a new function in utils to
cut down on the size of the function a bit and add independent testability.

- allow bounds checking in move_rel

Use the new functionality to optionally check bounds in move_rel (by
passing along the checker spec to _move).

- check high bounds in jog, swallow errors

When we jog during labware calibration, it's possible to trigger hard
limits, most commonly in +z. We should catch and prevent these, since
hard limits mess up the positioning state. But we have no way to
actually display errors during jog actions, so we just swallow the error
at the jog level. So, if you jog too far, the jog just doesn't do
anything.

Closes #6562