refactor(app): add cal information to pipette readiness selectors
Alters the (monster) selector getProtocolPipettesInfo to also include
whether or not a pipette offset calibration exists for the detected
pipette, which will later be used by UI components to determine whether
the protocol is ready to run. These selectors build in the feature flag
query, and will always indicate that calibration is present if the
feature flag for calibration overhaul is turned off (and now show up as
places to change when we look for things to remove along with the
feature flag later).

In addition to protocolPipettesMatch, which operates unchanged, we now
have protocolPipettesCalibrated (similar semantics but for offset
calibration, keeping in mind that the checks are based on the feature
flag) and protocolPipettesReady, which is Match && Calibrated. Ready is
used in the nav selectors to enable gating - since that's the point of
this exercise - but Match is still used in fileinfo, and Calibrated soon
will be as well.