refactor(app): s/hasOffsetCalibration/needsOffsetCalibration
Change the name and invert the sense of the prop that checks for pipette
calibration.

When the prop is named hasOffsetCalibration, it's a statement of fact
about the state of the system, and overriding it to `true` in some
circumstances makes that not actually true.

When the prop is named needsOffsetCalibration (and is inverted relative
to hasOffsetCalibration), it's a statement about the requirements that
must be met before the protocol can run, and it's a much more natural
fit for a prop that in reality depends on more information than just whether a
calibration exists for the pipette - information like, "is this pipette
actually used in the protocol" or "is the calibration overhaul feature
flag engaged".

This then makes it easier to consider whether a pipette is actually used
in a protocol when deciding whether it needs calibration.