refactor(app): robot settings calibration card (#6649)
When the calibration overhaul feature flag is selected, display a new Robot Calibration card in the robot settings. This card is the home for deck calibration; calibration check; and a display of pipette calibration data. When the feature flag is selected, the deck calibration and calibration check controls are displayed under it, and the new card is responsible for polling robot calibration status and data.

The only other content change is to display the date of the last deck calibration in the deck calibration control rather than a link to download deck calibration data.

Closes #6611