refactor(app): fetch pipettes when we view a robot in calibrationcard
This adds an initial fetch of pipettes and pipette offset calibrations
in the calibration card when a robot name chaanges or a robot changes
viewable state. Fixes an issue where the calibration card would display
no pipettes on when the user first selected a robot until they clicked
the pipettes & modules tab.