refactor(app,robot-server): Display pipette offsets in calibration card (#6664)
In the calibration card, we can now display pipette offset calibrations
as well as deck calibrations. These are pulled from the fusion of
attached pipettes and known pipette offset calibrations. Included in this
data is the URI of the tiprack used to execute the calibration, which was
always stored but was not previously sent via the HTTP interface (this is
the robot-server part of this commit).

We can use the tiprack uri and fuse it with our custom labware state
to lookup the name of the tiprack used to calibrate the pipette, and
also display the date of calibration.

Pipettes and pipette offset calibrations are fetched
in the calibration card when a robot name chaanges or a robot changes
viewable state. Fixes an issue where the calibration card would display
no pipettes on when the user first selected a robot until they clicked
the pipettes & modules tab.

Closes #6612