refactor(app): Map tip length cal data to TLC control (#6713)
We want the main button in the pre-protocol flow that starts tip length
calibration to be dynamic and say that you're going to recalibrate if
you have existing tip length data for the selected tiprack. There's a
helpful prop that does this, but it was being driven by the old probed
prop of the rpc pipettes. We want it to be driven by the existence of
TLC data for this tiprack + pipette combo.

To accomplish this, we need to change the pipettes component to actually
map in data from the http interface in addition to the rpc interface,
and then parse out the rpc pipettes to pass to other components. This
happens now.

The button now correctly says recalibrate or calibrate.

Closes #6695