refactor(app): use InlineCalibrationWarning
We're going to use the DeckCalibrationWarning in a couple other places
so let's make it a bit more generic.

Part of that is moving the logic about when to display it up to the
caller and keeping only the logic about how to render (color, content,
etc) in the component itself. While we're at it, choose how it renders
in the DeckCalibrationControl based on that control's knowledge of the
calibration data and status.