refactor(robot-server): Allow TLC recalibration with less data
When the user attaches a pipette, we want to do pipette offset
calibration. To do pipette offset calibration, you may or may not want
to specify a tiprack. If there is no pipette offset calibration stored
for the mount/pipette combo the user just attached, then we need to
figure out which tiprack to use, and depending on which tiprack you use,
you may want to do tip length calibration for it.

You can't figure out from available data which tip rack to use for
pipette offset calibration unless either
1. you force the user to tell you
2. you have the same lookup table for default labware in the client as
in the server

To fix this hole, this commit changes the semantics of pipette offset
calibration's createParams. The tip rack definition is now evaluated
before tip length recalibration (which is a new name) and the value of
tip length recalibration is based in part on the tip rack.

Specifically,
- If you specify a tiprack definition, it will be used for this pipette
offset calibration. If that specified tiprack definition does not have a
tip length calibration, recalibrateTipLength is forced to true;
otherwise, it is respected (so you can force a recalibrate where one
otherwise would not happen, but you cannot force POC to go ahead without
a tip length calibration)
- If you do not specify a tiprack definition, then a default one will be
generated: either the same one in the current pipette offset calibration
will be used, or (if there isn't a pipette offset calibration) the
default Opentrons tiprack for the volume is used. recalibrateTipLength
is respected or forced-on in the exact same way based on the exact same
criteria as if you did specify a tiprack definition.

The upshot of this is that clients in an attach-pipette scenario can
simply
- determine if a pipette offset calibration exists for this
mount/pipette combo
- if it does not, start a pipette offset calibration session without
specifying a tiprack or whether there should be a recalibration

and it will do the right thing.